simple render ways with line

+++ listref.h
@@ -1,12 +1,13 @@
 #include <stdlib.h>
 
+
 typedef struct st_listref listref;
 
 struct st_listref{
   listref *next;
-  int ref;
+  long ref;
 };
 
-listref *listref_create(int ref);
+listref *listref_create(long ref);
 
-listref *listref_append(listref *list, int ref);
+listref *listref_append(listref *list, long ref);
